Add StopOngoingVoiceBroadcastUseCase
